+++ test/encoder/pretty_printer_test.dart
@@ -345,6 +345,11 @@ void main() {
             prettyPrinter.visitBasicString(TomlBasicString('line 1\nline 2'));
             expect(prettyPrinter.toString(), equals(r'"line 1\nline 2"'));
           });
+          test('escapes Windows newlines', () {
+            var prettyPrinter = TomlPrettyPrinter();
+            prettyPrinter.visitBasicString(TomlBasicString('line 1\r\nline 2'));
+            expect(prettyPrinter.toString(), equals(r'"line 1\r\nline 2"'));
+          });
         });
         group('visitLiteralString', () {
           test('encodes simple literal strings correctly', () {
@@ -395,12 +400,22 @@ void main() {
               equals('"""\nfoo ""\\"""\\""bar""\\"""\\"" baz"""'),
             );
           });
-          test('does not escapes newlines', () {
+          test('does not escape newlines', () {
             var prettyPrinter = TomlPrettyPrinter();
             prettyPrinter.visitMultilineBasicString(
               TomlMultilineBasicString('line 1\nline 2'),
             );
             expect(prettyPrinter.toString(), equals('"""\nline 1\nline 2"""'));
+          });
+          test('does not escape Windows newlines', () {
+            var prettyPrinter = TomlPrettyPrinter();
+            prettyPrinter.visitMultilineBasicString(
+              TomlMultilineBasicString('line 1\r\nline 2'),
+            );
+            expect(
+              prettyPrinter.toString(),
+              equals('"""\nline 1\r\nline 2"""'),
+            );
           });
           test('escapes backslashes', () {
             var prettyPrinter = TomlPrettyPrinter();
@@ -437,12 +452,22 @@ void main() {
             );
             expect(prettyPrinter.toString(), equals("'''\nfoo'''"));
           });
-          test('does not escapes newlines', () {
+          test('does not escape newlines', () {
             var prettyPrinter = TomlPrettyPrinter();
             prettyPrinter.visitMultilineLiteralString(
               TomlMultilineLiteralString('line 1\nline 2'),
             );
             expect(prettyPrinter.toString(), equals("'''\nline 1\nline 2'''"));
+          });
+          test('does not escape Windows newlines', () {
+            var prettyPrinter = TomlPrettyPrinter();
+            prettyPrinter.visitMultilineLiteralString(
+              TomlMultilineLiteralString('line 1\r\nline 2'),
+            );
+            expect(
+              prettyPrinter.toString(),
+              equals("'''\nline 1\r\nline 2'''"),
+            );
           });
         });
       });